styled.js from MovieDetailsTile moved to common\StylesDetailsTile
